pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "setuptools_scm[toml]>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "entropy-table"
dynamic = ["version"]
description = "Contract-first relational entropy atlas engine for stochastic thermodynamics & open quantum systems"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.17",
]

[project.optional-dependencies]
dev = [
    "typer[all]>=0.12",
    "ruff>=0.6",
    "black>=24.8",
    "mypy>=1.11",
    "pre-commit>=4.0",
    "pytest>=8.3",
]

[project.scripts]
entropy-table = "entropy_table.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
write_to = "src/entropy_table/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
